+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- Windows line endings in macro definition are parsed correctly
+
 ## [1.1.4] - 2023-03-04
 
 ### Fixed